requirements.txt
streamlit
tensorflow==2.15.0
scikit-learn
pillow
numpy
joblib
